NlmeData

Documentation for NlmeData.

This package provides the data sets from the nlme package for R as Arrow files.

There are only two functions, datasets and dataset, defined, but not exported, in this package. The reason they are not exported is so that the names do not conflict with similar names in other packages.

NlmeData.dataset — Method

dataset(nm)

Return, as an Arrow.Table, the data set named nm, which can be a String or Symbol

NlmeData.datasets — Method

datasets()

Return a vector of names of the available data sets

The capitalization of the names of the datasets in the R package is preserved, even though it is more common to use lower case names in Julia.

using NlmeData: dataset, datasets
datasets()    # names of available data sets

48-element Vector{String}:
 "Alfalfa"
 "Assay"
 "BodyWeight"
 "CO2"
 "Cefamandole"
 "ChickWeight"
 "DNase"
 "Dialyzer"
 "Earthquake"
 "Fatigue"
 ⋮
 "Spruce"
 "Tetracycline1"
 "Tetracycline2"
 "Theophylline"
 "Wafer"
 "Wheat"
 "Wheat2"
 "bdf"
 "ergoStool"

An individual dataset always has a response variable and a grouping factor. For example, in the Rail dataset on the travel times of ultrasonic head-waves in 6 different samples of railway rails

dataset(:Rail)

Arrow.Table with 18 rows, 2 columns, and schema:
 :rail    String
 :travel  Int8

the response is travel and the grouping factor is rail.

Frequently there are other covariates recorded. For example, in a longitudinal data set one of the covariates will represent the time of the observation.

dataset(:Orange)

Arrow.Table with 35 rows, 3 columns, and schema:
 :tree           String
 :age            Int16
 :circumference  Int16

A convenient way to view an individual dataset is to create a GroupedDataFrame with DataFrames.groupby. (DataFrames is not a dependency of NlmeData - so as to avoid a heavyweight dependency. You must load it separately.)

using DataFrames
groupby(DataFrame(dataset(:Orange)), :tree)

GroupedDataFrame with 5 groups based on key: tree

First Group (7 rows): tree = "1"

tree	age	circumference

String	Int16	Int16

1	1	118	30

2	1	484	58

3	1	664	87

4	1	1004	115

5	1	1231	120

6	1	1372	142

7	1	1582	145

⋮

Last Group (7 rows): tree = "5"

tree	age	circumference

String	Int16	Int16

1	5	118	30

2	5	484	49

3	5	664	81

4	5	1004	125

5	5	1231	142

6	5	1372	174

7	5	1582	177